Fix subprocess output vs. python buffering.
As per the previous commit comment, python buffering can cause
print output to come out after subprocess command output.
There are number of ways to fix that (StackOverflow has so many
and many duplicates), but since we're using /usr/bin/env anyways,
the -S option lets you specify additional command line parameters
to the invoked program, so this feels like the simplest solution.
Assuming it works on a Mac for Peter of course!